+++ examples/src/main/scala/io/github/rafafrdz/criteria4s/examples/BetweenExample.scala
@@ -1,10 +1,10 @@
 package io.github.rafafrdz.criteria4s.examples
 
 import io.github.rafafrdz.criteria4s.core._
+import io.github.rafafrdz.criteria4s.dialect.mongodb.MongoDB
 import io.github.rafafrdz.criteria4s.examples.datastores.{Postgres, WeirdDatastore}
 import io.github.rafafrdz.criteria4s.extensions._
 import io.github.rafafrdz.criteria4s.functions._
-import io.github.rafafrdz.criteria4s.mongodb.MongoDB
 
 object BetweenExample extends App {
   val simpleBetween: Criteria[Postgres]    = col[Postgres]("a").between(range(1, 10))
