+++ examples/src/main/scala/io/github/rafafrdz/criteria4s/examples/Example4.scala
@@ -1,10 +1,10 @@
 package io.github.rafafrdz.criteria4s.examples
 
 import io.github.rafafrdz.criteria4s.core._
+import io.github.rafafrdz.criteria4s.dialect.mongodb.MongoDB
 import io.github.rafafrdz.criteria4s.examples.datastores.{MySQL, Postgres, WeirdDatastore}
 import io.github.rafafrdz.criteria4s.extensions._
 import io.github.rafafrdz.criteria4s.functions._
-import io.github.rafafrdz.criteria4s.mongodb.MongoDB
 
 object Example4 extends App {
