+++ mongodb/src/main/scala/io/github/rafafrdz/criteria4s/dialect/mongodb/package.scala
@@ -1,4 +1,4 @@
-package io.github.rafafrdz.criteria4s
+package io.github.rafafrdz.criteria4s.dialect
 
 package object mongodb extends MongoDB.MongoDBExpr[MongoDB] {
   type MongoDBExpr[T <: MongoDB] = MongoDB.MongoDBExpr[T]
